+++ R/utils.R
@@ -19,6 +19,8 @@ vdapply <- function(X, FUN, ...) {
   vapply(X, FUN, FUN.VALUE = double(1), ...)
 }
 
+# Like mapply, but better recycling rules, and no simplifying
+
 mapx <- function(...) {
   args <- list(...)
   if (length(args) == 0) stop("No arguments to `mapx()`")
@@ -58,69 +60,18 @@ add_attr <- function(x, attr, value) {
   x
 }
 
-read.dcf.gz <- function(x) {
-  con <- gzfile(x, open = "r")
-  on.exit(close(con))
-  read.dcf(con)
-}
-
-get_cran_extension <- function(platform) {
-  switch(
-    platform,
-    "source" = ".tar.gz",
-    "macos" = ".tgz",
-    "windows" = ".zip",
-    stop("Unknown platform: ", sQuote(platform))
-  )
-}
-
 get_platform <- function() {
-  .Platform
+  R.version$platform
 }
 
-get_minor_r_version <- function(x) {
-  x <- package_version(x)
-  vapply(unclass(x), function(x) paste(x[1:2], collapse = "."), character(1))
-}
-
-get_all_package_dirs <- function(platforms, rversions) {
-  minors <- unique(get_minor_r_version(rversions))
-  if (any(package_version(minors) < "3.2")) {
-    stop("pkgcache does not support packages for R versions before R 3.2")
-  }
-  res <- lapply(platforms, function(pl) {
-    if (pl == "source") {
-      cbind("source", "*", "src/contrib")
-
-    } else if (pl == "windows") {
-      cbind("windows", minors, paste0("bin/windows/contrib/", minors))
-
-    } else if (pl == "macos") {
-      res1 <- lapply(minors, function(v) {
-        pv <- package_version(v)
-        if (pv >= "4.0") {
-          cbind("macos", v, paste0("bin/macosx/contrib/", v))
-        } else if (package_version(v) <= "3.3") {
-          cbind("macos", v, paste0("bin/macosx/mavericks/contrib/", v))
-        } else {
-          cbind("macos", v, paste0("bin/macosx/el-capitan/contrib/", v))
-        }
-      })
-      do.call(rbind, res1)
-    }
-  })
-
-  mat <- do.call(rbind, res)
-  colnames(mat) <- c("platform", "rversion", "contriburl")
-  res <- as_tibble(mat)
-  res$prefix <- paste0(
-    "/",
-    ifelse(res$rversion == "*", "*", paste0("R-", res$rversion)),
-    "/", res$platform, "/"
-  )
-
-  res
-}
+# Why are we using this?
+# AFAICT the only difference is that if `getOption("encoding")` is set,
+# then it is observed in `file()` and then `readLines()` converts the
+# file to UTF-8. (At least on R 4.1, earlier versions are probably
+# different.)
+#
+# OTOH it seems that we only use it to read the etag from a file, plus
+# in some test cases, so it should not matter much.
 
 read_lines <- function(con, ...) {
   if (is.character(con)) {
@@ -161,18 +112,20 @@ interpret_dependencies <- function(dp) {
 ## so we would need an R version specific vector here.
 ## Not an issue currently, might be in the future.
 
-#' @importFrom utils installed.packages
-
 base_packages <- function() {
   if (is.null(repoman_data$base_packages)) {
     repoman_data$base_packages <-
-      rownames(installed.packages(priority = "base"))
+      parse_installed(.Library, priority="base")$Package
   }
   repoman_data$base_packages
 }
 
 is_na_scalar <- function(x) {
-  length(x) == 1 && is.na(x)
+  identical(x, NA_character_) ||
+    identical(x, NA_integer_) ||
+    identical(x, NA_real_) ||
+    identical(x, NA_complex_) ||
+    identical(x, NA)
 }
 
 drop_nulls <- function(x)  {
@@ -228,4 +181,27 @@ modify_vec <- function(old, new) {
 
 last <- function(x) {
   x[[length(x)]]
+}
+
+get_os_type <- function() {
+  .Platform$OS.type
+}
+
+encode_path <- function(path) {
+  if (get_os_type() == "windows") {
+    enc2utf8(path)
+  } else {
+    enc2native(path)
+  }
+}
+
+gzip_decompress <- function(from, chunk_size = 5 * 1000 * 1000) {
+  con <- gzcon(rawConnection(from))
+  on.exit(close(con), add = TRUE)
+  pieces <- list()
+  while (1) {
+    pieces[[length(pieces) + 1]] <- readBin(con, what = "raw", n = chunk_size)
+    if (length(pieces[[length(pieces)]]) == 0) break;
+  }
+  do.call("c", pieces)
 }